code cleanup

+++ bot.php
@@ -8,11 +8,11 @@ use Discord\Parts\Channel\Channel;
 use Discord\Parts\Channel\Message;
 
 $token = $_ENV['TOKEN'];
-$GLOBALS["wold_channel_id"] = $_ENV['WORLD_CHANNEL_ID'];
-$GLOBALS["wold_channel_mdl"] = $_ENV['WORLD_CHANNEL_MESSAGE_DAYS_LIMIT'];
-$GLOBALS["wold_channel_mcl"] = $_ENV['WORLD_CHANNEL_MESSAGE_COUNT_LIMIT'];
-$GLOBALS["wold_channel_throttle"] = $_ENV['WORLD_CHANNEL_THROTTLE'];
-$GLOBALS['last_throttle_time'] = 0;
+$GLOBALS["channel_id"] = $_ENV['WORLD_CHANNEL_ID'];
+$GLOBALS["channel_mdl"] = $_ENV['WORLD_CHANNEL_MESSAGE_DAYS_LIMIT'];
+$GLOBALS["channel_mcl"] = $_ENV['WORLD_CHANNEL_MESSAGE_COUNT_LIMIT'];
+$GLOBALS["channel_clean_throttle"] = $_ENV['WORLD_CHANNEL_THROTTLE'];
+$GLOBALS['channel_last_clean_throttle'] = 0;
 
 $discord = new Discord([
     'token' => $token,
@@ -32,38 +32,21 @@ $discord->on('init', function (Discord $discord) {
             return;
         }
 
-        if ($message->channel_id !== $GLOBALS["wold_channel_id"]) {
+        if ($message->channel_id !== $GLOBALS["channel_id"]) {
             return;
         }
 
         if ($message->content === '!config') {
-            $message->channel->sendMessage("WORLD_CHANNEL_ID: " . $GLOBALS["wold_channel_id"]);
-            $message->channel->sendMessage("WORLD_CHANNEL_MESSAGE_DAYS_LIMIT: " . $GLOBALS["wold_channel_mdl"]);
-            $message->channel->sendMessage("WORLD_CHANNEL_MESSAGE_COUNT_LIMIT: " . $GLOBALS["wold_channel_mcl"]);
+            $message->channel->sendMessage("WORLD_CHANNEL_ID: " . $GLOBALS["channel_id"]);
+            $message->channel->sendMessage("WORLD_CHANNEL_MESSAGE_DAYS_LIMIT: " . $GLOBALS["channel_mdl"]);
+            $message->channel->sendMessage("WORLD_CHANNEL_MESSAGE_COUNT_LIMIT: " . $GLOBALS["channel_mcl"]);
             return;
         }
 
         reactToBumi($message);
-
-        if (!hasBeenThrottled()) {
-            return;
-        };
-
         cleanChannelMessages($message->channel);
     });
 });
-
-function hasBeenThrottled()
-{
-    // Throttle: Solo permite la acción si han pasado X segundos desde la última ejecución
-    $currentTime = time(); // Tiempo actual en segundos
-    if ($currentTime - $GLOBALS['last_throttle_time'] < $GLOBALS["wold_channel_throttle"]) {
-        return false;
-    }
-
-    $GLOBALS['last_throttle_time'] = $currentTime;
-    return true;
-}
 
 function reactToBumi(Message $message)
 {
@@ -75,12 +58,17 @@ function reactToBumi(Message $message)
 
 function cleanChannelMessages(Channel $channel)
 {
+    $currentTime = time();
+    if ($currentTime - $GLOBALS['channel_last_clean_throttle'] < $GLOBALS["channel_clean_throttle"]) {
+        return false;
+    }
+    $GLOBALS['channel_last_clean_throttle'] = $currentTime;
     echo '[THROTTLE RESET] LIMPIANDO MENSAJES - ' . time() . PHP_EOL;
 
     // delete messages older than X days
     $channel->getMessageHistory(['limit' => 100])->then(function ($messages) use ($channel) {
         $deletables = $messages->filter(function ($message) {
-            $days = 60 * 60 * 24 * $GLOBALS["wold_channel_mdl"];
+            $days = 60 * 60 * 24 * $GLOBALS["channel_mdl"];
             return time() - $message->timestamp->getTimestamp() > $days;
         });
         $channel->deleteMessages($deletables);
@@ -88,12 +76,12 @@ function cleanChannelMessages(Channel $channel)
 
     // delete messages over the limit
     $channel->getMessageHistory(['limit' => 100])->then(function ($messages) use ($channel) {
-        if ($messages->count() > $GLOBALS["wold_channel_mcl"]) {
+        if ($messages->count() > $GLOBALS["channel_mcl"]) {
             $deletables = Collection::for(Message::class);
             $index = 0;
             foreach ($messages as $message) {
                 $index++;
-                if ($index > $GLOBALS["wold_channel_mcl"]) {
+                if ($index > $GLOBALS["channel_mcl"]) {
                     $deletables->pushItem($message);
                 }
             }
